hl7-sender/requirements.txt
hl7apy==1.3.4
requests==2.28.2
python-dotenv==1.0.0
faker==18.11.2 